misra: make hal MISRA-compliant for all architectures
JIRA: DO-432, DO-431, DO-433, DO-434 DO-435, DO-436, DO-437, DO-438

+++ hal/arm/scs.h
@@ -21,19 +21,19 @@
 #include "hal/types.h"
 
 
-void _hal_scsIRQSet(u32 irqn, u8 state);
+void _hal_scsIRQSet(u8 irqn, u8 state);
 
 
-void _hal_scsIRQPrioritySet(u32 irqn, u32 priority);
+void _hal_scsIRQPrioritySet(u8 irqn, u32 priority);
 
 
-void _hal_scsIRQPendingSet(u32 irqn);
+void _hal_scsIRQPendingSet(u8 irqn);
 
 
-int _hal_scsIRQPendingGet(u32 irqn);
+int _hal_scsIRQPendingGet(u8 irqn);
 
 
-int _hal_scsIRQActiveGet(u32 irqn);
+int _hal_scsIRQActiveGet(u8 irqn);
 
 
 void _hal_scsPriorityGroupingSet(u32 group);
