+++ posix/inet.c
@@ -50,7 +50,7 @@ static ssize_t sockcall(unsigned int socket, msg_t *msg)
 		return err;
 	}
 
-	err = smo->ret;
+	err = (int)smo->ret;
 	return err;
 }
 
@@ -110,7 +110,7 @@ int inet_accept4(unsigned int socket, struct sockaddr *address, socklen_t *addre
 		return err;
 	}
 
-	return err;
+	return (int)err;
 }
 
 
@@ -121,7 +121,7 @@ int inet_bind(unsigned int socket, const struct sockaddr *address, socklen_t add
 	hal_memset(&msg, 0, sizeof(msg));
 	msg.type = sockmBind;
 
-	return sockdestcall(socket, &msg, address, address_len);
+	return (int)sockdestcall(socket, &msg, address, address_len);
 }
 
 
@@ -132,7 +132,7 @@ int inet_connect(unsigned int socket, const struct sockaddr *address, socklen_t 
 	hal_memset(&msg, 0, sizeof(msg));
 	msg.type = sockmConnect;
 
-	return sockdestcall(socket, &msg, address, address_len);
+	return (int)sockdestcall(socket, &msg, address, address_len);
 }
 
 
@@ -143,7 +143,7 @@ int inet_getpeername(unsigned int socket, struct sockaddr *address, socklen_t *a
 	hal_memset(&msg, 0, sizeof(msg));
 	msg.type = sockmGetPeerName;
 
-	return socknamecall(socket, &msg, address, address_len);
+	return (int)socknamecall(socket, &msg, address, address_len);
 }
 
 
@@ -154,7 +154,7 @@ int inet_getsockname(unsigned int socket, struct sockaddr *address, socklen_t *a
 	hal_memset(&msg, 0, sizeof(msg));
 	msg.type = sockmGetSockName;
 
-	return socknamecall(socket, &msg, address, address_len);
+	return (int)socknamecall(socket, &msg, address, address_len);
 }
 
 
@@ -174,7 +174,7 @@ int inet_getsockopt(unsigned int socket, int level, int optname, void *optval, s
 	ret = sockcall(socket, &msg);
 
 	if (ret < 0) {
-		return ret;
+		return (int)ret;
 	}
 
 	*optlen = (socklen_t)ret;
@@ -191,7 +191,7 @@ int inet_listen(unsigned int socket, int backlog)
 	msg.type = sockmListen;
 	smi->listen.backlog = backlog;
 
-	return sockcall(socket, &msg);
+	return (int)sockcall(socket, &msg);
 }
 
 
@@ -304,7 +304,7 @@ int inet_shutdown(unsigned int socket, int how)
 	msg.type = sockmShutdown;
 	smi->send.flags = (unsigned int)how;
 
-	return sockcall(socket, &msg);
+	return (int)sockcall(socket, &msg);
 }
 
 
@@ -320,7 +320,7 @@ int inet_setsockopt(unsigned int socket, int level, int optname, const void *opt
 	msg.i.data = optval;
 	msg.i.size = optlen;
 
-	return sockcall(socket, &msg);
+	return (int)sockcall(socket, &msg);
 }
 
 
@@ -333,7 +333,7 @@ int inet_setfl(unsigned int socket, unsigned int flags)
 	msg.type = sockmSetFl;
 	smi->send.flags = (unsigned int)flags;
 
-	return sockcall(socket, &msg);
+	return (int)sockcall(socket, &msg);
 }
 
 
@@ -344,5 +344,5 @@ int inet_getfl(unsigned int socket)
 	hal_memset(&msg, 0, sizeof(msg));
 	msg.type = sockmGetFl;
 
-	return sockcall(socket, &msg);
+	return (int)sockcall(socket, &msg);
 }
